requirements.txt
django
psycopg2-binary
python-decouple
djangorestframework
djangorestframework-simplejwt
drf_yasg
openpyxl

